Bug 1748718 - Part 2: Use the specified type for rust xpidl constants, r=emilio
Previously all xpidl constants were specified as `i64` which means they require
casts before being passed to any xpcom methods. The lack of typing was not an
issue in c++ due to implicit casts from enums to integer types, but using the
correct type is much more valuable in Rust.

Differential Revision: https://phabricator.services.mozilla.com/D135165

UltraBlame original commit: 9d8cfa10fa95c3f9287abe7d426ebbebd0cc4747

+++ toolkit/components/glean/api/src/ipc.rs
@@ -103,7 +103,7 @@ pub fn need_ipc() -> bool {
 
 #[cfg(feature = "with_gecko")]
 fn register_process_shutdown(process_type: u32) {
-    match process_type as i64 {
+    match process_type {
         nsIXULRuntime::PROCESS_TYPE_DEFAULT => {
             
         }
